engine: Refresh LUN using host from different DC to scan
When trying to execute the refresh LUN command the engine uses
all the hosts in the system to scan all the selected LUNs,
instead of LUN per host scan.

Due to that "Object currently in use" error raises,
since one or more hosts try to scan the same LUN.

This fix separates the scanned LUNs into groups by
the host that is attached to and performs the scan
LUN operation for each group separately.

Change-Id: I6d41aa73115dc68efef0a6f71d56e5e81dfe0f14
Bug-Url: https://bugzilla.redhat.com/1838051
Signed-off-by: Bella Khizgiyev <bkhizgiy@redhat.com>