packaging: setup: Improve customization texts
Add an empty line before more-than-one-line prompts, to make it easier
to see where a question begins.

Split (for ovn, for now) the text to two: An explanation, then, on a new
line, the actual short question. IMO this makes it easier to see what
you are actually asked about.

Change-Id: Ic5ab976b0d0bcfac1520f9998e6a6dfef5b8f10c
Signed-off-by: Yedidyah Bar David <didi@redhat.com>