core: fix UpdateVmCommand for UpdateClusterCommand
When upgrading from 4.3 to 4.4, the user initiate the
UpdateClusterCommand. This, invokes UpdateVmCommand on each VM in the
cluster and uses compensation context for rolling back. For emulated
machine or chipset changes, it was not implemented. Therefore, throwing
an exception.

This patch implements the compensation support for emulated machine and
chipset changes, letting the UpdateClusterCommand succeed with those
changes to the VMs.

Change-Id: Ia9e4bbbf7d6d6219fa03c792e4b47aaa5812b556
Bug-Url: https://bugzilla.redhat.com/1847513
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>