engine: Add changedFields to next-run Snapshots
The UI calculates the changed fields between current and next run
configuration on each refresh. This causes performance issues on big
setups (especially after cluster update, when all VMs have next run
config snapshot created).

This calculation is not necessary, it is possible to calculate it just
once (when creating snapshot) and use this value to show to the user.

This patch adds a new table column changed_fields to the snapshots table
that contains the fields that were changed between current and next run
configuration.

Change-Id: I6f8c792fdc312f33e62d969a487ab39115c5dbcd
Bug-Url: https://bugzilla.redhat.com/1845747
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>