core: update new root checkpoint after checkpoint removal
When the root checkpoint removed from the chain, the next
checkpoint in the chain becomes the new root, it means
that the XML of that checkpoint should be updated and also
the checkpoint parent is now NULL.

The updated checkpoint XML should be fetched from the host by
calling the new checkpoint_xml API after the parent (previous root)
was removed.

VM.checkpoint_xml -
{
  'vmID': vm-id,
  'checkpoint_id': checkpoint-id
}

Response -
{
  'result': {
    'checkpoint': <checkpoint-xml>
  }
}

This patch adds the update of the new root checkpoint.

Change-Id: I58f1e1ec63f3f25a968af6fd0d725d671ac80054
Bug-Url: https://bugzilla.redhat.com/1813831
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>