packaging: setup: Drop note with a link on separate machine PKI
It was added for websocket-proxy, when it was only manual.
Now, if using ssh as root, it's trivial. Even if using "manual files",
it's rather easy.

The link for websocket-proxy does not exist (used to exist on the old
mediawiki-based website). The link for grafana points to a page I
created just for this code, currently just a stub.

Change-Id: I28e83a7ac96938df64081d75da83b5b0e277f177
Signed-off-by: Yedidyah Bar David <didi@redhat.com>