core: backward compatbility to old devices alias
When having a VM running in old engine versions, the aliases were set as
libvirt report them. If the user didn't shutdown the VM and upgrade, he
will able to suffer from devices being unplugged.

This patch brings back the devices correlation for old aliases when the
devices are managed.

Change-Id: Iccc330f2d6bc3474323434e16a97bbe71f10edea
Bug-Url: https://bugzilla.redhat.com/1850220
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>