webadmin: mark all tabs as valid before validation
Problem: validation state of SYSTEM_TAB was not cleared
at the start of validation cycle. Once the tab was marked as invalid
it stayed as such even if problem was corrected by the user.
Fix: in order to prevent similar cases in the future method for clearing
the validation state of all tabs was added.

Change-Id: I160cc0a352ae992039fdb7d7c2258d5b14eceac9
Bug-Url: https://bugzilla.redhat.com/1820208
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>