core: Start NBD server in readonly mode for download
When downloading disks using NBD backend, use readonly=true. Starting
read-write NBD server for download is bad and may lead to corruption of
existing volume.

This change fixes templates download. Currently starting template download
from the SDK fail in vdsm, since we send readonly=False, and vdsm does
not allow starting NBD server in read-write mode for templates:

    vdsm.storage.exception.SharedVolumeNonWritable: Shared volume is
    read only: ['sdUUID: 56ecc03c-4bb5-4792-8971-3c51ea924d2e',
    'imgUUID: 9b62b5fa-920e-4d0c-baf6-40406106e48e',  'volUUID:
    ed6dc087-632c-40d9-9b0d-9ccc7dcce0e8']

When download fails, the transfer remains in "initializing" phase, the
disk remain locked, and volume on the host remain active. There is no
way to clean up since cancelling the transfer is disabled. User need to
modify the database manually and deactivate the volume on the host.

This change avoids the error when starting the NBD server, but more work
is needed to handle errors properly when starting the server fails.

Change-Id: Id247963005e6bae1378f5b6f4afc60fc611f6f4e
Bug-Url: https://bugzilla.redhat.com/1851708
Signed-off-by: Nir Soffer <nsoffer@redhat.com>