core: Enable image transfer cancellation
ImageTransfer.cancel() was disabled in:

commit 52c57cef573be0b1d31a9e2c2370288fd351aa4a
Author: Ahmad Khiet <akhiet@redhat.com>
Date:   Mon May 4 16:32:57 2020 +0300

    webadmin: cancel image upload for webadmin only

That change broke basic functionality - being able to cancel a transfer
when using the API. Canceling a transfer is required for cleanup after
upload image failures, or failures during image transfer initialization
or finalization.

The change needed was to disable cancel option in the UI for image
transfer started using the API. Instead this change disabled cancelling
for transfer started from the API.

Cancelling a transfer from the API fails now with this this error:

Traceback (most recent call last):
  File "./cancel_transfer.py", line 109, in <module>
    connection, types.ImageTransfer(id=args.transfer_id))
  File "/home/nsoffer/src/ovirt-engine-sdk/sdk/examples/helpers/imagetransfer.py",
line 226, in cancel_transfer
    transfer_service.cancel()
  File "/usr/local/lib64/python3.7/site-packages/ovirtsdk4/services.py",
line 13748, in cancel
    return self._internal_action(action, 'cancel', None, headers, query, wait)
  File "/usr/local/lib64/python3.7/site-packages/ovirtsdk4/service.py",
line 299, in _internal_action
    return future.wait() if wait else future
  File "/usr/local/lib64/python3.7/site-packages/ovirtsdk4/service.py",
line 55, in wait
    return self._code(response)
  File "/usr/local/lib64/python3.7/site-packages/ovirtsdk4/service.py",
line 296, in callback
    self._check_fault(response)
  File "/usr/local/lib64/python3.7/site-packages/ovirtsdk4/service.py",
line 132, in _check_fault
    self._raise_error(response, body)
  File "/usr/local/lib64/python3.7/site-packages/ovirtsdk4/service.py",
line 118, in _raise_error
    raise error
ovirtsdk4.Error: Fault reason is "Operation Failed". Fault detail is
"[Cannot cancel image transfer from any source other than the
browser.]". HTTP response code is 409.

See this patch for example image transfer cancellation:
https://gerrit.ovirt.org/c/109917/

Reverting the change is not trivial because of the database scripts, and
Daniel started to work on a better fix here:
https://gerrit.ovirt.org/c/109923/

But this require more work, and we need to restore the functionality.
This change is a minimal temporary fix restoring the previous behavior.

Change-Id: I1fd208f9923a7cf2648a1d008b9051ac0679e8b7
Bug-Url: https://bugzilla.redhat.com/1661816
Signed-off-by: Nir Soffer <nsoffer@redhat.com>