core: persist VM configuration on rerun
Previously, we lost some of the VM configuration when executing rerun.
Since the command already initiated, we clear the cached VM and proceed
to validate and execute phases. We would have configuration that wasn't
been re-initialized.

This patch will call init method on rerun flow to make sure we
re-initalizing the VM configuration, making it persist between the rerun
attempts.

Change-Id: I305613c329a4ada12c44874f6999ff66a13973ad
Bug-Url: https://bugzilla.redhat.com/1807937
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>