db: Update cluster bios type for current clusters with CL <= 4.3
oVirt 4.4 added a new bios_type field to the cluster, with the
default value set to Cluster Default (0). However this creates an
issue for existing clusters, since for them the value that is
supposed to be updated on the addition of the first host never
happens.

This patch takes care of this, and sets the value to Legacy (1),
for all existing cluster of oVirt 4.3 and lower.

Change-Id: Iff305e80ce30cebaf19266e49bb4ace29c6a35a2
Signed-off-by: Lev Veyde <lveyde@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1852040