engine: Fix add managed vm validation
In AddVmCommand there was a correct validation for the kubevirt vm but
it was not called because the vms created via UI did not have origin
type set.

Instead of relying on the vm's origin we can check the templates
isManaged method.

Also, the error message dispayed to the user was changed from
"The specified disk does not exist." to
"No disks have been specified".

Change-Id: I1b5672d177d756dbb3718cfe2095547faa09dba4
Bug-Url: https://bugzilla.redhat.com/1845396
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>