core: Export VM with collapse snapshots to destination SD
When trying to export a VM from a storage domain to another, the
export command runs successfully, but the new disks of the
exported VM are created on the source SD and not on the destination.
The issue happens only when the collapse snapshot option is selected.

This patch fixes the issue and creates the disk on the right SD.

Change-Id: I33fe806ae8393f49203ecea7fc65fb9bd41bd4c6
Bug-Url: https://bugzilla.redhat.com/1842272
Signed-off-by: Bella Khizgiyev <bkhizgiy@redhat.com>