core: Don't default COW Block to prealloc
When importing images with COW file format to a Block SD,
the allocation type was being set to Preallocated which
is not supported because the libvirt sparseness implemenation
does not currently support Block Storage.

Changing the VDSM to support COW format for image files for
Block Storage uses the directio implementation which now allows
for these scenarios to be supported when the allocation type
is set to thin allocation or sparse.

The code for defaulting the allocation type to Preallocated
for COW image file format when the destination is Block Storage
has been removed. Also, for importing COW formated files with
a sparse allocation type to Block Storage, the actual size of
the image is set. The warning for applying virtual size to
Block Storage imports with KVM has been modified so that
it is limited to raw files.

Now we can import COW image files to Block Storage when sparse
allocation type is set.

Change-Id: I1035cfd1c5af58f2c84cbf00d4e2d99ddae302ce
Bug-Url: https://bugzilla.redhat.com/1663135
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>