engine: Add VM to KubeVirt refactoring
When creating a KubeVirt VM from the oVirt UI, the AddVmCommand
has been called 2 times:
1. by the UI to invoke KubeVirt API
2. by the VmUpdater to insert the new KubeVirt into oVirt DB

This was confusing and required several conditions for checking if the
vm is managed and if it is called as internal action.

This patch creates a new action for calling the KubeVirt API called
AddVmToKubevirtCommand. It makes it easier to see the requirements for
calling the API (mainly cluster, template and instance disks).

When the newly created KubeVrt VM is discovered by the monitoring, the
correct VMStatic and AddVMParams are created. That should allow the
AddVmCommand to be unaware if it is managed or non managed VM.

Change-Id: Ifeec3000cc797672e9e3cb03abefa0a126feaf7b
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>