core: run global maintenance on hosted engine's run_on_vds
When triggering global maintenance operation from the webadmin,
we executed SetHaMaintenanceMode on the selected host. This is
problematic because during upgrade, one may trigger it on a host
that is configured for the non-active hosted engine (the one of
the lower compatibility version).

This is a quick fix for that: the webadmin no longer sends the
host to execute the operation on to the backend. Instead, the
backend finds the host that the hosted engine VM runs on and
execute the SetHaMaintenanceMode on it.

From REST-API the backend gets the host that the hosted engine
VM runs on, so no need to change anything in that flow. Ideally,
the webadmin should also do that operation on the hosted engine
VM (rather than on a host) and the validation of the VM should
move from the REST-API to the backend.

Change-Id: I1616485b4f4c756b884e3d3fee09c62d19720de5
Bur-Url: https://bugzilla.redhat.com/1853225
Signed-off-by: Arik Hadas <ahadas@redhat.com>