core: fix race on VdsManager::lastVmsList
When the VM is powering up, the engine has pending resources on the host
it's going to run. When the VM is switching to PoweringUp status, these
resources are released and should be taken from the host monitoring.
It is possible for a small period of time that the host wouldn't catch
that we have the VM running and won't calculate its resources, letting
us overcommit to the host which can lead us to qemu proceess OOM.

The main reason for the problematic VdsManager::lastVmsList value which
holds the cached vm resources, is that only PollVmStatsRefresher::poll
is setting it after querying the VMs on that host. The thread runs every
15 seconds. In the meanwhile HostMonitoring::refreshCommitedMemory can
be called, and it uses the lastVmsList value in order to calculate the
host resources, overriding the right values.

As a result, when getting the event of VM is powering up, we can add it
to lastVmsList ensuring the real time status of lastVmsList value.

Change-Id: If771b25751e3cfb1f153055667cea12ad9f68590
Bug-Url: https://bugzilla.redhat.com/1692355
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>