core: safely remove ovirt guest agent channel
Since we can't know much about the OS installed in the guest, the safest
way to remove the ovirt guest agent channel is when the user indicates
he has OS that doesn't need that channel. We know it should be removed
for el8 and above.

This patch will remove the unnecessary ovirt-guest-agent channel from
the VM on creation, based on the OS Type selected. For now it's only for
el8 and OSes that derive from it.

Change-Id: Icf5008d6f4f72ccca10f5098b5d26c3059d42a77
Bug-Url: https://bugzilla.redhat.com/1845591
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>