WebAdmin: Move the Remove cluster button to the menu
Users can sometime be confused between the button
to remove the Cluster within a detailed tab view,
and the remove button for removing an item with the
tab detailed view. Such as when viewing a Virtual
Machine within the Virtual Machines tab of a Cluster
within the Cluster screen. This means that the user
can delete the Cluster itself when they thought they
were just removing a sub item.

To make the system more user friendly, the remove
button has been moved to the kebana drop down menu
so that this mistake is mitigated. This change should
allow for greater usability.

Change-Id: I10ee9ec10eee5c0693e36aa5c3e45032260f13d7
Bug-Url: https://bugzilla.redhat.com/1667019
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>