core: NUMA changes to be applied on Next Run
Modifying the NUMA Node pinning was not being handled
via the Next Run and not being updated when the Virtual
Machine is running.

This fix sets the Numa Node list edit to support
Next Run and ensures that the modified Numa List is
saved and also adds support for Next Run cpu pinning.

Change-Id: Ie2abe02fc1f2d3572b3b331f1534f558e91a7975
Bug-Url: https://bugzilla.redhat.com/1573218
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>