fix null in diskattachment herf
When attaching a disk to a VM, the returned disk-attachment object
contains 'null' as part of the href string:

  <distattachment href="/ovirt-engine/api/vms/xxx/null/yyy">

This is due to a rare case in the API in which an entity (disk_attachment)
has the same parent (Vm) in more than one place in the API tree:

  .../api/vms/yyy/diskattachments
  .../api/storagedomains/xxx/vms/yyy/diskattachments

This causes ambiguity in the link-generation process. The assumption that
the pair <entity, parent> leads to a unique location in the API tree is
strongly embedded in the code and is not worth changing due to 1 or 2
exceptional cases. Therefore, the solution is hard-coded

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1842457
Change-Id: Ib53eb79c557465051f52488f8d6f4e87d4cb7b9f
Signed-off-by: Ori_Liel <oliel@redhat.com>