webadmin: distinguish KubeVirt entities
KubeVirt entities should be distinguished from the normal ones in the
WebAdmin GUI.

This patch adds a new icon (inspired by PF) that represents the
container. The icon is used in status columns for VM, VM templates,
hosts, clusters, storage domains  and data centers.

The template list did not have status column so far and since it would
be used only for OKD/OpenShift templates, the column is hidden by
default.

Change-Id: I224899830fe6e741422e960249ba6f93a6ac6021
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>