packaging: pki: renew: Check if ca file exists
When renewing PKI, check if ca file exists, before checking if it
expired.

In 4.4, we added a new ca file for qemu, and now check both of them for
expiry. On upgrade from 4.3, this ca file does not exist yet, at this
point. So if we need to renew during upgrade from 4.3, without this
patch, we fail, when checking if it expired.

Change-Id: I58795266c47440cdb6950c8ab5db7b45bf848df8
Bug-Url: https://bugzilla.redhat.com/1841195
Signed-off-by: Yedidyah Bar David <didi@redhat.com>