core: change GA sign to QEMU
This patch will consider the QEMU guest agent status to sign it on the
VM. Since we have virtio-win, a new logic and comparison should be
executed. When the old windows guest tools are installed, the QEMU guest
agent doesn't have version. Therefore, checking if it exists on the VM
is a good indicator that we need to update it.

Change-Id: Iaad726a2fd9bef1f4c1b11f3833ac5bee9e5e9ff
Bug-Url: https://bugzilla.redhat.com/1828347
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>