core: Prevent powering on VM during disk migration
When performing a disk migration from one SD to another
the VM can't be powered on.

This patch adds validation for the case that the disk is locked
or the storage pool is down for HA VMs due to disk/storage migration,
powering on operation will be blocked until the operation is finished.

Change-Id: Iac9a086fdb30063c7a31e5aaec49a6a02325faeb
Bug-Url: https://bugzilla.redhat.com/1840732
Signed-off-by: Bella Khizgiyev <bkhizgiy@redhat.com>