webadmin: import and upgrade storage domain fix
Importing and upgrading a storage domain
from V4 to V5 will result in a failure
in the UI and the storage domain will
not imported.

The root cause for the NPE is that when
the upgrade message is presented, the
list of the storage domains to add is
set to null, this list is used in the
later methods.

This patch fixes the UI error and fixes
importing the storage domain failure.

Screenshots: https://imgur.com/a/fowf0fF

Change-Id: Ib8c855697c94c2ba17b82039a7a87a94ccf33aaa
Bug-Url: https://bugzilla.redhat.com/1828089
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>