core: disable callback for PVCDisks
Creating PVCDisk would trigger a callback like a regular DiskImage did,
despite the fact they only require database operations. `useCallback`
was supposed to prevent this, but AddDiskCommand#createDiskBasedOnImage
overrode the parameters with a new DiskImage object, making the callback
run regardless.

This patch skips the override for disks that are not regular images.

Change-Id: I2db03df82fd9ecc40ea2d1f5afb6882510e2a4d5
Bug-Url: https://bugzilla.redhat.com/1855377
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>