core: Add 'custom' attribute to BiosType element in OVF
If user has explicitly set the BIOS type of the VM, add custom="true"
attribute to <BiosType> element when exporting the VM to OVF. Negative
value of this attribute or its absence means that BIOS type value was
derived by the VM from the cluster default.

On import, if a VM/template uses the cluster default BIOS type
(custom="false" or not set in <BiosType>), OvfReader sets
VmBase.biosType field to CLUSTER_DEFAULT. The actual BIOS type of the
cluster the VM/template comes from is stored in the separate field
clusterBiosTypeOrigin of VM or VmTemplate class respectively.

Change-Id: Ie487e8b59a7ac682ee305e6abd6d6ea81fa18f48
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>