webadmin: set target domain for multiple disks move/copy
screenshot: https://imgur.com/a/peIq0Ot

Currently, while moving/copying multiple disks, the user is required
to set the target domain for each disk separately, one by one.

This RFE allows setting the same target domain for multiple disks at
once on the move/copy popup.
The common target domains list defined as the intersection of the
available target domains of each disk.
In case of setting a different target domain for part of the disks,
so not all disks will be moved/copied to the same one,
the common target domain would set to 'Mixed'.

Change-Id: Ib0d19e8d0fed6d88631339780ba246e890002552
Bug-Url: https://bugzilla.redhat.com/1749803
Signed-off-by: Shani Leviim <sleviim@redhat.com>