cnv: prometheus failures logging improvements
This patch changes the way errors are handled when  connecting to prometheus or
fetching metrics.

Each specific error is logged only once in WARN category. Continues
logging can be enabled by enabling DEBUG category.
An error can be logged again only when the particular problem is fixed (successfully re-connected /
metrics fetched) or when the engine is restarted.

Change-Id: Ib4914af3ea6d70f22d17438fca20a8e6d6538d61
Bug-Url: https://bugzilla.redhat.com/1845432
Signed-off-by: Artur Socha <asocha@redhat.com>