engine: Add config to override disk propagation errors
In OVirt, we have a property propagate_error at the disk level
that decides in case of an error, how this error be propagated
to the VM. This value is maintained in the database table with
the default value set as Off. The default setting(Off) results
in a policy that ends up pausing the VM. If you want to overide
the behavior, there is no provision in the UI currently to
configure this property for disk and thus there is no easy way
to set this value.
Setting the value for propagate_erro to "Off" is not ideal for
disks of type LUN and Cinder. We need to provide a way for
customers to decide how to handle the out of the box behavior.
This change introduces a new engine config for customers to set
the propagate_error property to be On if desired. This will
result in disks of types LUN and Cinder to report the errors to
VM and let them handle it rather than stopping them. This also
avoids setting of the flag disk-by-disk and provides clean way
for customers to change the behavior reliably.
To summarize again, the out of the box behavior for LUN and
Cinder disks is not changed but additional capability is
provided for these disk type to propagate errors if desired.

Change-Id: Id87f42d32b7b047375f37acbb864fa1a12e4bbad
Bug-Url: https://bugzilla.redhat.com/1862534
Signed-off-by: Shubha Kulkarni <shubha.kulkarni@oracle.com>