webadmin+vdsbroker: disable TSC in 4.3 cluster
TSC introduced in 4.4, it requires the VDSM to report the TSC in the
host caps. But, when we have 4.4 engine with 4.3 hosts or 4.3 cluster we
will have the value as null. In that case, we should skip writing the
TSC clock to the domxml, preventing NPE when running VMs.
In addition the option to manually set it in the UI should be
unchangeable and set to false in 4.3 clusters.

Change-Id: Ie673b39178885853422101c36b8c773b70eaaf8e
Bug-Url: https://bugzilla.redhat.com/1863615
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>