core: change snapshot disks status on snapshot removal
Currently, while removing a snapshot, its disks haven't been locked
and their status remains 'OK'.
This patch changes their status to 'LOCKED' on their snapshot removal.

In case of injecting an error on LSM, which causes an image to an
ILLEGAL status, all other images on the chain should be unlocked.

Change-Id: Ib2b668e36ab2919f3afa2fa11c0bb0ad86e32381
Bug-Url: https://bugzilla.redhat.com/1828241
Signed-off-by: Shani Leviim <sleviim@redhat.com>