core: change the way address is assigned to hot-plugged disks
Previously, on every hot-plug/unplug operation we had passed the devices
to 'getDiskAddressMap' method that passed it to 'getAddressMapForScsiDisk'
that set the address of the given device internally. It was easy to miss
where the address is set with this approach.

Now we do it differently:
1. We rather return the address back from 'getDiskAddressMap', therefore
it is renamed to 'getDiskAddress'.
2. The address is now set in 'performPlugCommand' and only in plug flow
as the address is not used in unplug flow.

This patch also includes some cleanup in 'getAddressMapForScsiDisk'.

Change-Id: Ia70834c15434b514ba9b01d94f208d525c06cba1
Bug-Url: https://bugzilla.redhat.com/1855305
Signed-off-by: Arik Hadas <ahadas@redhat.com>