core: remove CreateImageVDSCommand
Both CreateImageVDSCommand and CreateVolumeVDSCommand are using the same
VDSM API call - createVolume(), the only difference is that
CreateVolumeVDSCommand sets the following members:
 - imageId
 - sourceImageGroupId

The above fields set to Guid.Empty in CreateImageVDSCommand.

This patch removes the code duplication by using the
CreateVolumeVDSCommand and setting the extra fields to Guid.Empty,
CreateVolumeVDSCommandParameters and CreateImageVDSCommandParameters
are now merged to the same object.

Change-Id: Ib763206fe744c495be82fa3d76c776d2c3a4ac5f
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>