core: propagate address instead of vmdevice
For the calculation of an address to disks, we only need their existing
address and their interface. Now that we set the address of the device
in AbstractDiskCommand#performPlugCommand, there is no reason to pass
the whole VmDevice instance down the flow and we can propagate just the
device address.

Change-Id: Ib4bc8b9a27c96d84e6e75ee23a4318dd8f647223
Bug-Url: https://bugzilla.redhat.com/1855305
Signed-off-by: Arik Hadas <ahadas@redhat.com>