core: Add the ability to switch hosted engine storage to master
In the case of two SD in the system when one of them is hosted-storage
and the second is the master, if the master SD being sent to
maintenance the hosted-storage doesn't pick up the master role
automatically and all the DC becomes inactive.

This patch fixes the issue and the hosted-storage
can became the master.

Change-Id: I44edd256734a67fb6ad1093010fe425c8cf3c733
Bug-Url: https://bugzilla.redhat.com/1836034
Signed-off-by: Bella Khizgiyev <bkhizgiy@redhat.com>