core: support adding bitmaps in cold snapshot operation
If a cold snapshot is taken for a VM that contains backups,
the new 'active' volume will not contain the created backups
bitmaps.

The engine should instruct the host to copy the bitmaps to
the new active (top) volume. A new flag was added for that
reason to createVolume API verb.

The new addBitmaps flag will be set to 'true' only if the
snapshot is taken while the VM is down and only if the disk
is part of a backup.

Change-Id: I0263f69eba0e773b0f3552edc40a893f950c0d09
Bug-Url: https://bugzilla.redhat.com/1861667
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>