core: Validate iso file when not KVM
When importing a Virtual Machine from a KVM server
when an iso file was uploaded to disk storage and
an iso domain is not available, the import would
fail because an iso file was detected, but an iso
domain was not.

Added validation so that only when the external server
is not a KVM server do we validate that when there is
an iso file uploaded to disk, that the iso domain
exists.

Now the KVM import succeeds when an iso file is applied
outside of an iso domain and the iso domain has not
been imported nor created.

Change-Id: I52bba4dd55e7cbafb361b7fee6762c374c054b00
Bug-Url: https://bugzilla.redhat.com/1632068
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>