ansible: use mode preserve when copying to host
There were some changes in ansible 2.9.12 where when you use module copy it
automatically gives the file 0600 instead of original mode.
This also was happening to files that were created by parameter content.
We need to specify the type because it does not have original file.

Updated permissions and ownership in /etc/pki/vdsm to match the previous state.
Removed unnecessary cleanup of hosted_engine_conf.

Change-Id: Iecf4871ccb6f34127610dc8df6f2101704bbe98e
Bug-Url: https://bugzilla.redhat.com/1869298
Signed-off-by: Martin Necas <mnecas@redhat.com>