core+restapi: introduce AutoPinningPolicy
This patch will introduce the enum AutoPinningPolicy. This enum will
indicate how we wish to set the CPU and NUMA pinning automatically for a
VM.
The enum has two options:
1. `existing`, this will set the CPU and NUMA pinning to the VM without
   changing the CPU topology of it.
2. `adjust`, this will change the VMs CPU topology to fix and maximize the
   host it is pinned on, it will automatically set the CPU and NUMA
   pinning accordingly.

The current usage is only when adding a VM, it must have a dedicated
host.

Having the enum set, it will generate and seta CPU pinning string
accroding the VM and host CPU topology. It will also pin the NUMA nodes
setting the numa policy to strict. This will get the best possible
performance to that VM with that VM and host CPU topoligies.

Change-Id: I10d7871d9f9d02df2cc37809b0aa399f061969d2
Bug-Url: https://bugzilla.redhat.com/1862968
Signed-off-by: Liran Rotenberg <lrotenbe@redhat.com>