search: force query to use tag view
When searching for VMs with certain cluster & host properties the
generated sql uses vms.cluster_name instead of
vms_with_tags.cluster_name adding a redundant huge nested join (in scaled env)
of the vms view instead using the vms_with_tags view that is already
there.

The result is a long running query that runs several minutes

This patch addresses this issue by forcing using the vms_with_tags table
when certain conditions are matched (as the problematic query)

The side affect of this change is improvement of two other queries
represented in the tests where vms.cluster_name is replaced by vms_with_tags.cluster_name

A new test was added as well to check the original problematic search
string.

This patch is safe and AFAICT will not cause any regressions.

Change-Id: Id08fac4b195122c7965f8103b04fdadb3c1a882a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1858638
Signed-off-by: Eli Mesika <emesika@redhat.com>