webadmin: resume silent image upload fix
Pausing an image upload, and resuming the
upload process back throws NullPointerException
error on the UI, without the ability to resume
the upload back.

The root cause of this bug is the silent upload
process, which is being checked if it can be done.

There's no need to add the Storage Domain id in
the parameters, the disk id is the only parameter
that is needed for silent upload.

Screenshots: https://imgur.com/a/IYosqrx

Change-Id: I7ed75d60018e5698607498b5c2e9d6d46c8fe022
Bug-Url: https://bugzilla.redhat.com/1828333
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>