core: Fix stopping transfer session
If removing proxy ticket fails, for example when using development
environment without running ovirt-imageio, we return early from
stopImageTransferSession() without stopping the NBD server.

In this case image verification will fail, since qemu-nbd takes an
exclusive lock on the volume, and qemu-img info will fail to take a
lock:

    qemu-img: Could not open [/path/to/image]: Failed to get shared "write" lock
    Is another process using the image [/path/to/image]?

The flow is this:

- Stopping transfer session
- Remove host ticket succeeds
- Remove proxy ticket fails
- We return true because this transfer is not done via a browser
- Continue normal cleanup
- Fail to verify image

The is a regression introduced by:

commit 67e836b538d7fc79fea0560d865733de325ce6ea
Author: Daniel Erez <derez@redhat.com>
Date:   Wed Jul 22 10:46:07 2020 +0300

    core: avoid duplicate code in TransferDiskImage

Before that change, we stopped the NBD server when removing proxy ticket
failed, and the transfer was not started from the UI.

Fixed by always ignoring failures when removing proxy ticket.  Once we
remove the host ticket, the image cannot be accessed by anyone, so
having a stale proxy ticket is harmless. The only possible failure is
that the proxy is not running, and in this case there is nothing to
remove, and there is no point to retry the operation.

With this change the flow is:

- Stopping transfer session
- Remove host ticket succeeds
- Remove proxy ticket fails (logging the error)
- Stop NBD server
- Verify image (succeeds)

Change-Id: Id3e4acb6ef132334281a78a087be176436342b56
Bug-Url: https://bugzilla.redhat.com/1872383
Signed-off-by: Nir Soffer <nsoffer@redhat.com>