core: Support disabling imageio proxy
Some users are still using the legacy unsupported All-In-One[1]
deployment. In this configuration, both imageio proxy (port 54323) and
imageio daemon (port 54322) are used on engine host. Since ovirt-imageio
2.0 ovirt-imageio does not support such configuration.

However since we run imageio daemon on engine host, we don't need
the imageio proxy, and we can work directly with the daemon.

Add a new config value: ImageTransferProxyEnabled, set to true by
default.  Users that still use All-In-One deployment can disable this
option using:

    engine-config -s "ImageTransferProxyEnabled=false"
    systemctl restart ovirt-engine

When the proxy is disabled:

- Engine does not add, remove, or extend proxy image tickets.
- ImageTransfer's proxy_url is same as transfer_url, so the change is
  transparent to applications using proxy_url.
- Transferring images from administration portal accesses ovirt-imageio
  daemon (port 54322) directly.

[1] https://www.ovirt.org/develop/release-management/features/integration/allinone.html

Change-Id: Id7fc21f0529f6d294288bed7a5e3873d01e886ce
Bug-Url: https://bugzilla.redhat.com/1871348
Signed-off-by: Nir Soffer <nsoffer@redhat.com>