core: make sure a hot-plugged disk is assigned with unique address
Previously, we didn't take the address of host SCSI devices that
are passthrough to a VM into account when allocating an address
for a SCSI disk. This could have resulted in a conflict between
a SCSI disk and a passthrough SCSI disk that are both assigned
with the same address.

This patch addresses this by loading the assigned addresses of
host SCSI devices that the VM is set with when allocating an
address for SCSI disks. This way, when looking for the next
available address to assign to a SCSI disk, we skip addresses
that are already used by VM devices that correspond to host
SCSI devices.

Change-Id: I89770e43e5484f3afe3ef6170a552fec01068afe
Bug-Url: https://bugzilla.redhat.com/1855305
Signed-off-by: Arik Hadas <ahadas@redhat.com>