core: recursive search for quota
Clients should specify the quota to consume in data centers in
which quota is enabled. However, that assumption already breaked
with the VM portal that expects the engine to pick an arbitrary
quota the user can consume, just to be able to trigger an
operation. While this is wrong and buggy, it does not seem to be
cost-effective to change this now.

So this patch changes the lookup for an arbitrary quota to
use to also include quotas that are inherited from the user's
parent elements.

Change-Id: Ic970cb75c8b5d3c3e83a7c4b8c44065beb387839
Bug-Url: https://bugzilla.redhat.com/1822372
Signed-off-by: Arik Hadas <ahadas@redhat.com>