core: allocate unused macs before used ones
A MAC pool might have free MAC addresses in the pool but they may be used
elsewhere on plugged vnics in the system.
Currently the in-pool allocation mechanism allocates addresses blindly
without checking whether they are used elsewhere.

This patch changes the behavior to allocate globally unused addresses
as long as globally unused addresses are available and continue to allocate
addresses which are not used in the pool, but used globally.

Change-Id: Ie6c24cb0f0737f2bba609aff384c79569dc27877
Bug-Url: https://bugzilla.redhat.com/1760170
Signed-off-by: Eitan Raviv <eraviv@redhat.com>