packaging: Add transactions helpers module
Add transactions helper module providing RemoveFileTransaction. This can
be used to remove a file during upgrade as part of the upgrade
transaction.

I think this can be part of otopi, since this is a pretty generic
helper, but we can move it to otopi later. This can also replace
ExternalTruststoreRemoveTransaction.

Change-Id: I2f84695cc934c1ee65004bc9320157d56c9ad21c
Bug-Url: https://bugzilla.redhat.com/1871348
Signed-off-by: Nir Soffer <nsoffer@redhat.com>