webadmin: persist grid settings on the server
Summary:
1. chosen grid settings:
  - hiding/showing grid columns
  - swapping grid columns
2. feature enabled by default
3. feature can be disabled via Options popup (user can use the grids
   as before)
4. settings are stored in DB as part of user.options column (JSON)
   under key 'webAdmin'
5. existing settings are migrated (uploaded) from local storage on
   first login

Grids and local storage

Grids persist several settings in the local storage (as key/value).
This has been extended to save chosen keys also on the server.
The local storage continues to be "the single source of truth" - all
settings are read directly from there however right now it is
populated on user log-in with content stored on the server.
Later each local change (of the supported keys) is mirrored
to the server. Each upload takes a full snapshot (all keys currently
in local storage).

Data transfer

On log-in the settings are injected into the web admin page and
available via logged-in user instance. When UI is fully operational,
the GWT RPC is used to fetch/upload user data. Under the hood the same
UpdateUserSettingsCommand is used as by REST API.
Currently delivery is best effort only - there is no retry
in case of network errors. Changed but not uploaded settings will
continue to work locally until next log in.

Change-Id: Ifb6463b6574ab659d5ccc848f553ca0f919bf0ff
Bug-Url: https://bugzilla.redhat.com/1752751
Signed-off-by: Radoslaw Szwajkowski <rszwajko@redhat.com>